✨ refactor: Moved helpers.
