+++ src/utils/helpers.ts
@@ -11,14 +11,3 @@ export function sleep(ms: any) {
 export function capitalize(str: string) {
 	return str.charAt(0).toUpperCase() + str.slice(1);
 }
-
-export const morganProdOptions =
-	'":remote-addr/:remote-user" [:date[web]] ' +
-	'":method :url HTTP/:http-version" :status ' +
-	'":referrer" ":user-agent" :response-time ms';
-
-export const corsOptions = {
-	methods: ['GET', 'HEAD', 'PUT', 'PATCH', 'POST', 'DELETE'],
-	preflightContinue: true,
-	optionsSuccessStatus: 200,
-};
